refactor(api): move update campaigns to its bounded context
