+++ api/lib/application/campaigns/campaign-management-controller.js
@@ -2,7 +2,6 @@ import { usecases } from '../../domain/usecases/index.js';
 import { extractParameters } from '../../infrastructure/utils/query-params-utils.js';
 import * as campaignDetailsManagementSerializer from '../../infrastructure/serializers/jsonapi/campaign-details-management-serializer.js';
 import * as participationForCampaignManagementSerializer from '../../infrastructure/serializers/jsonapi/participation-for-campaign-management-serializer.js';
-import { deserializer } from '../../infrastructure/serializers/jsonapi/deserializer.js';
 
 const getCampaignDetails = async function (request) {
   const campaignId = request.params.id;
@@ -22,21 +21,9 @@ const findPaginatedParticipationsForCampaignManagement = async function (request
   return participationForCampaignManagementSerializer.serialize(participationsForCampaignManagement, meta);
 };
 
-const updateCampaignDetailsManagement = async function (request, h) {
-  const campaignId = request.params.id;
-
-  const campaignDetailsManagement = await deserializer.deserialize(request.payload);
-  await usecases.updateCampaignDetailsManagement({
-    campaignId,
-    ...campaignDetailsManagement,
-  });
-  return h.response({}).code(204);
-};
-
 const campaignManagementController = {
   getCampaignDetails,
   findPaginatedParticipationsForCampaignManagement,
-  updateCampaignDetailsManagement,
 };
 
 export { campaignManagementController };
