+++ api/src/prescription/campaign/infrastructure/repositories/campaign-management-repository.js
@@ -1,9 +1,8 @@
-import _ from 'lodash';
-import { knex } from '../../../db/knex-database-connection.js';
-import { CampaignManagement } from '../../domain/read-models/CampaignManagement.js';
-import { fetchPage } from '../utils/knex-utils.js';
-import { CampaignParticipationStatuses } from '../../../src/prescription/shared/domain/constants.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { CampaignManagement } from '../../../../../lib/domain/read-models/CampaignManagement.js';
+import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
+import { CampaignParticipationStatuses } from '../../../shared/domain/constants.js';
+import { CampaignTypes } from '../../domain/read-models/CampaignTypes.js';
 
 const { SHARED, TO_SHARE, STARTED } = CampaignParticipationStatuses;
 
@@ -78,20 +77,7 @@ const findPaginatedCampaignManagements = async function ({ organizationId, page 
   return { models: campaignManagement, meta: { ...pagination } };
 };
 
-const update = function ({ campaignId, campaignAttributes }) {
-  const editableAttributes = _.pick(campaignAttributes, [
-    'name',
-    'title',
-    'customLandingPageText',
-    'customResultPageText',
-    'customResultPageButtonText',
-    'customResultPageButtonUrl',
-    'multipleSendings',
-  ]);
-  return knex('campaigns').where({ id: campaignId }).update(editableAttributes);
-};
-
-export { get, findPaginatedCampaignManagements, update };
+export { get, findPaginatedCampaignManagements };
 
 async function _countParticipationsByStatus(campaignId, campaignType) {
   const row = await knex('campaign-participations')
