+++ api/tests/acceptance/application/campaigns/campaign-management-controller_test.js
@@ -1,4 +1,4 @@
-import { databaseBuilder, expect, generateValidRequestAuthorizationHeader, knex } from '../../../test-helper.js';
+import { databaseBuilder, expect, generateValidRequestAuthorizationHeader } from '../../../test-helper.js';
 import { createServer } from '../../../../server.js';
 
 describe('Acceptance | API | Campaign Management Controller', function () {
@@ -46,40 +46,6 @@ describe('Acceptance | API | Campaign Management Controller', function () {
       // then
       expect(response.statusCode).to.equal(200);
       expect(response.result.data[0].id).to.equal(campaignParticipation.id.toString());
-    });
-  });
-
-  describe('PATCH /api/admin/campaigns/{id}', function () {
-    it('should return the updated campaign', async function () {
-      // given
-      const campaign = databaseBuilder.factory.buildCampaign({ name: 'odlName', multipleSendings: false });
-      const user = databaseBuilder.factory.buildUser.withRole();
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject({
-        method: 'PATCH',
-        url: `/api/admin/campaigns/${campaign.id}`,
-        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
-        payload: {
-          data: {
-            attributes: {
-              name: 'newName',
-              title: campaign.title,
-              'custom-landing-page-text': campaign.customLandingPageText,
-              'custom-result-page-button-text': null,
-              'custom-result-page-button-url': null,
-              'custom-result-page-text': null,
-              'multiple-sendings': true,
-            },
-          },
-        },
-      });
-      const updatedCampaign = await knex('campaigns').first();
-      // then
-      expect(response.statusCode).to.equal(204);
-      expect(updatedCampaign.name).to.equal('newName');
-      expect(updatedCampaign.multipleSendings).to.be.true;
     });
   });
 });
