+++ api/tests/prescription/campaign/acceptance/application/campaign-administration-route_test.js
@@ -4,6 +4,7 @@ import {
   generateValidRequestAuthorizationHeader,
   learningContentBuilder,
   mockLearningContent,
+  knex,
 } from '../../../../test-helper.js';
 
 import { ORGANIZATION_FEATURE } from '../../../../../lib/domain/constants.js';
@@ -301,6 +302,40 @@ describe('Acceptance | API | campaign-administration-route', function () {
 
       const response = await server.inject(options);
       expect(response.statusCode).to.equal(204);
+    });
+  });
+
+  describe('PATCH /api/admin/campaigns/{id}', function () {
+    it('should return the updated campaign', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ name: 'odlName', multipleSendings: false });
+      const user = databaseBuilder.factory.buildUser.withRole({ role: ROLES.SUPER_ADMIN });
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'PATCH',
+        url: `/api/admin/campaigns/${campaign.id}`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
+        payload: {
+          data: {
+            attributes: {
+              name: 'newName',
+              title: campaign.title,
+              'custom-landing-page-text': campaign.customLandingPageText,
+              'custom-result-page-button-text': null,
+              'custom-result-page-button-url': null,
+              'custom-result-page-text': null,
+              'multiple-sendings': true,
+            },
+          },
+        },
+      });
+      const updatedCampaign = await knex('campaigns').first();
+      // then
+      expect(response.statusCode).to.equal(204);
+      expect(updatedCampaign.name).to.equal('newName');
+      expect(updatedCampaign.multipleSendings).to.be.true;
     });
   });
 });
