+++ api/tests/unit/application/campaign/index_test.js
@@ -38,169 +38,6 @@ describe('Unit | Application | Router | campaign-router ', function () {
     });
   });
 
-  describe('PATCH /api/admin/campaigns/{id}', function () {
-    it('should return 204', async function () {
-      // given
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport').callsFake((request, h) => h.response(true));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-      sinon
-        .stub(campaignManagementController, 'updateCampaignDetailsManagement')
-        .callsFake((request, h) => h.response('ok').code(204));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-      const payload = {
-        data: {
-          type: 'campaigns',
-          attributes: {
-            name: 'name',
-            title: 'title',
-            'custom-landing-page-text': null,
-            'custom-result-page-text': null,
-            'custom-result-page-button-text': null,
-            'custom-result-page-button-url': null,
-            'multiple-sendings': false,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/campaigns/1', payload);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-    });
-
-    it('should return 400 with an invalid campaign id', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/campaigns/invalid');
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-
-    it('should return 400 when name is null', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-      const payload = {
-        data: {
-          type: 'campaigns',
-          attributes: {
-            name: null,
-            title: null,
-            'custom-result-page-text': null,
-            'custom-result-page-button-text': null,
-            'custom-result-page-button-url': null,
-            'custom-landing-page-text': null,
-            'multiple-sendings': false,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/campaigns/1', payload);
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-
-    it('should return 400 when name is empty', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-      const payload = {
-        data: {
-          type: 'campaigns',
-          attributes: {
-            name: '',
-            title: null,
-            'custom-result-page-text': null,
-            'custom-result-page-button-text': null,
-            'custom-result-page-button-url': null,
-            'custom-landing-page-text': null,
-            'multiple-sendings': false,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/campaigns/1', payload);
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-
-    it('should return 400 when title is empty', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-      const payload = {
-        data: {
-          type: 'campaigns',
-          attributes: {
-            name: 'name',
-            title: '',
-            'custom-result-page-text': null,
-            'custom-result-page-button-text': null,
-            'custom-result-page-button-url': null,
-            'custom-landing-page-text': null,
-            'multiple-sendings': false,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/campaigns/1', payload);
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-
-    it('returns forbidden access if admin member has CERTIF role', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif').callsFake((request, h) => h.response(true));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-      const payload = {
-        data: {
-          type: 'campaigns',
-          attributes: {
-            name: 'name',
-            title: 'title',
-            'custom-landing-page-text': null,
-            'custom-result-page-text': null,
-            'custom-result-page-button-text': null,
-            'custom-result-page-button-url': null,
-            'multiple-sendings': false,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/campaigns/1', payload);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-  });
-
   describe('GET /api/admin/campaigns/{id}', function () {
     it('should return 200', async function () {
       // given
